backend/requirements.txt
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.1
Django==3.2.5
graphene-django==2.15.0
django-cors-headers==3.6.0
django-tinymce==3.3.0
djangorestframework == 3.12.4